# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IgwebuIke Alumni Network API
  version: 2.3.0
  description: Secure alumni networking API. Authentication uses an httpOnly cookie named `iw_token`.
servers:
- url: https://igwebuike-backend-1.onrender.com
- url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: iw_token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
  responses:
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/jobs:
    get:
      summary: 'List published jobs (public). Filters: q (text search), type, skill, location, page, limit<=20'
      responses:
        '200':
          description: OK
    post:
      summary: Post a job. Members' jobs are PENDING until an admin approves; employers/admins publish immediately
      responses:
        '200':
          description: OK
        '401': &id001
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: Created
        '400':
          description: Validation failed
      security: &id002
      - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - title
              - company
              - location
              - type
              - description
              properties:
                title:
                  type: string
                company:
                  type: string
                location:
                  type: string
                type:
                  type: string
                  enum:
                  - full-time
                  - part-time
                  - contract
                  - internship
                  - remote
                description:
                  type: string
                  minLength: 30
                  maxLength: 5000
                skills:
                  type: array
                  items:
                    type: string
                  maxItems: 15
                applyUrl:
                  type: string
                  description: https only
  /api/jobs/mine:
    get:
      summary: My posted jobs with application counts
      responses:
        '200':
          description: OK
        '401': *id001
      security: *id002
  /api/jobs/{id}:
    get:
      summary: One published job (public)
      parameters: &id003
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-f0-9]{24}$
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /api/jobs/{id}/apply:
    post:
      summary: Apply to a job (once). Optional alumni referral code
      responses:
        '200':
          description: OK
        '401': *id001
        '201':
          description: Submitted
        '400':
          description: Invalid / own job
        '404':
          description: Not found
        '409':
          description: Already applied
      security: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - coverNote
              properties:
                coverNote:
                  type: string
                  minLength: 20
                  maxLength: 2000
                cvUrl:
                  type: string
                referrerCode:
                  type: string
                  example: IWU-CSC-20-4821
      parameters: *id003
  /api/jobs/{id}/applications:
    get:
      summary: Applicants of a job (owner or admin; others get 404)
      responses:
        '200':
          description: OK
        '401': *id001
        '404':
          description: Not found or not yours
      security: *id002
      parameters: *id003
  /api/jobs/{id}/close:
    patch:
      summary: Close a job (owner or admin)
      responses:
        '200':
          description: OK
        '401': *id001
        '404':
          description: Not found or not yours
      security: *id002
      parameters: *id003
...
  /api/jobs/recommended:
    get:
      summary: Jobs ranked for me. ML ranking if the matching service is up, else keyword fallback. Excludes my own and already-applied
        jobs.
      security: *id004
      responses:
        '200':
          description: '{ model, needsProfile, recommendations:[{rank, score(0..1), matched[], job}] }'
        '401': *id005
  /api/jobs/recommendations/events:
    post:
      summary: Log feedback on a recommendation (view | apply | dismiss)
      security: *id004
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - jobId
              - action
              - rank
              - model
              properties:
                jobId:
                  type: string
                action:
                  type: string
                  enum:
                  - view
                  - apply
                  - dismiss
                rank:
                  type: integer
                  minimum: 1
                  maximum: 50
                model:
                  type: string
                  maxLength: 60
      responses:
        '201':
          description: Recorded
        '400':
          description: Invalid event
        '401': *id005